pyproject.toml
[project]
name = "google-search-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
    "playwright>=1.58.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
    "pytesseract>=0.3.13",
]
